+++ bindgen-tests/tests/expectations/tests/class_with_typedef.rs
@@ -93,19 +93,19 @@ impl Default for C {
 impl C {
     #[inline]
     pub unsafe fn method(&mut self, c: C_MyInt) {
-        unsafe { C_method(self, c) }
+        C_method(self, c)
     }
     #[inline]
     pub unsafe fn methodRef(&mut self, c: *mut C_MyInt) {
-        unsafe { C_methodRef(self, c) }
+        C_methodRef(self, c)
     }
     #[inline]
     pub unsafe fn complexMethodRef(&mut self, c: *mut C_Lookup) {
-        unsafe { C_complexMethodRef(self, c) }
+        C_complexMethodRef(self, c)
     }
     #[inline]
     pub unsafe fn anotherMethod(&mut self, c: AnotherInt) {
-        unsafe { C_anotherMethod(self, c) }
+        C_anotherMethod(self, c)
     }
 }
 #[repr(C)]