+++ bindgen-tests/tests/expectations/tests/issue-410.rs
@@ -37,7 +37,7 @@ pub mod root {
         impl Value {
             #[inline]
             pub unsafe fn a(&mut self, arg1: root::JSWhyMagic) {
-                unsafe { Value_a(self, arg1) }
+                Value_a(self, arg1)
             }
         }
     }