+++ bindgen-tests/tests/expectations/tests/issue-807-opaque-types-methods-being-generated.rs
@@ -89,15 +89,13 @@ extern "C" {
 impl Opaque {
     #[inline]
     pub unsafe fn eleven_out_of_ten(&mut self) -> SuchWow {
-        unsafe { Opaque_eleven_out_of_ten(self) }
+        Opaque_eleven_out_of_ten(self)
     }
     #[inline]
     pub unsafe fn new(pup: Pupper) -> Self {
-        unsafe {
-            let mut __bindgen_tmp = ::std::mem::MaybeUninit::uninit();
-            Opaque_Opaque(__bindgen_tmp.as_mut_ptr(), pup);
-            __bindgen_tmp.assume_init()
-        }
+        let mut __bindgen_tmp = ::std::mem::MaybeUninit::uninit();
+        Opaque_Opaque(__bindgen_tmp.as_mut_ptr(), pup);
+        __bindgen_tmp.assume_init()
     }
 }
 extern "C" {