+++ bindgen-tests/tests/expectations/tests/union-in-ns_1_0.rs
@@ -16,11 +16,11 @@ pub mod root {
         }
         #[inline]
         pub unsafe fn as_ref(&self) -> &T {
-            unsafe { ::std::mem::transmute(self) }
+            ::std::mem::transmute(self)
         }
         #[inline]
         pub unsafe fn as_mut(&mut self) -> &mut T {
-            unsafe { ::std::mem::transmute(self) }
+            ::std::mem::transmute(self)
         }
     }
     impl<T> ::std::default::Default for __BindgenUnionField<T> {