+++ bindgen-tests/tests/expectations/tests/union_dtor.rs
@@ -62,6 +62,6 @@ impl Default for UnionWithDtor {
 impl UnionWithDtor {
     #[inline]
     pub unsafe fn destruct(&mut self) {
-        unsafe { UnionWithDtor_UnionWithDtor_destructor(self) }
+        UnionWithDtor_UnionWithDtor_destructor(self)
     }
 }