+++ bindgen-tests/tests/expectations/tests/wrap_unsafe_ops_dynamic_loading_simple.rs
@@ -10,21 +10,19 @@ pub struct TestLib {
     __library: ::libloading::Library,
     pub foo: Result<
         unsafe extern "C" fn(
+            x: ::std::os::raw::c_int,
+            y: ::std::os::raw::c_int,
+        ) -> ::std::os::raw::c_int,
+        ::libloading::Error,
+    >,
+    pub bar: Result<
+        unsafe extern "C" fn(
             x: *mut ::std::os::raw::c_void,
         ) -> ::std::os::raw::c_int,
         ::libloading::Error,
     >,
     pub baz: Result<
-        unsafe extern "C" fn(
-            x: *mut ::std::os::raw::c_void,
-        ) -> ::std::os::raw::c_int,
-        ::libloading::Error,
-    >,
-    pub bazz: Result<
-        unsafe extern "C" fn(
-            arg1: ::std::os::raw::c_int,
-            ...
-        ) -> ::std::os::raw::c_int,
+        unsafe extern "C" fn() -> ::std::os::raw::c_int,
         ::libloading::Error,
     >,
 }
@@ -34,7 +32,7 @@ impl TestLib {
         P: AsRef<::std::ffi::OsStr>,
     {
         let library = ::libloading::Library::new(path)?;
-        Self::from_library(library)
+        unsafe { Self::from_library(library) }
     }
     pub unsafe fn from_library<L>(
         library: L,
@@ -43,26 +41,34 @@ impl TestLib {
         L: Into<::libloading::Library>,
     {
         let __library = library.into();
-        let foo = __library.get(b"foo\0").map(|sym| *sym);
-        let baz = __library.get(b"baz\0").map(|sym| *sym);
-        let bazz = __library.get(b"bazz\0").map(|sym| *sym);
+        let foo = unsafe { __library.get(b"foo\0") }.map(|sym| *sym);
+        let bar = unsafe { __library.get(b"bar\0") }.map(|sym| *sym);
+        let baz = unsafe { __library.get(b"baz\0") }.map(|sym| *sym);
         Ok(TestLib {
             __library,
             foo,
+            bar,
             baz,
-            bazz,
         })
     }
     pub unsafe fn foo(
         &self,
-        x: *mut ::std::os::raw::c_void,
+        x: ::std::os::raw::c_int,
+        y: ::std::os::raw::c_int,
     ) -> ::std::os::raw::c_int {
-        (self.foo.as_ref().expect("Expected function, got error."))(x)
+        unsafe {
+            (self.foo.as_ref().expect("Expected function, got error."))(x, y)
+        }
     }
-    pub unsafe fn baz(
+    pub unsafe fn bar(
         &self,
         x: *mut ::std::os::raw::c_void,
     ) -> ::std::os::raw::c_int {
-        (self.baz.as_ref().expect("Expected function, got error."))(x)
+        unsafe {
+            (self.bar.as_ref().expect("Expected function, got error."))(x)
+        }
+    }
+    pub unsafe fn baz(&self) -> ::std::os::raw::c_int {
+        unsafe { (self.baz.as_ref().expect("Expected function, got error."))() }
     }
 }