+++ bindgen-tests/tests/headers/wrap_unsafe_ops_objc_class.h
@@ -1,0 +1,10 @@
+// bindgen-flags: --wrap-unsafe-ops -- -x objective-c
+// bindgen-osx-only
+
+@class Foo;
+
+Foo* fooVar;
+
+@interface Foo
+-(void)method;
+@end